HorizonZero/requirements/dev.txt
-r base.txt
pytest
pytest-django
coverage
black
isort
flake8